fix: do not redirect to / on lang change
Now stays on the same page when changing language.

The essence of the issue: Django sends POST to change a language. To
stay on the same page it uses http header `referer`. Nginx maps
avarus.space to ip in local network but doesn't change referer header,
therefore Django get header with avarus.space in it. And because Django
knows nothing about avarus.space it redirects to /.

Used solution is override referer header in nginx.